+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.2.0 (unreleased)
 ------------------
 
+- #1911 Converted Container to Dexterity Contents
 - #1931 Removed archetypes.schemaextender from senaite.core
 - #1953 Allow custom id formatting regardless of portal type
 - #1952 Open analysis specification ranges